[Refactor]: 사람 수 Long 대신 int 로 타입 변경

+++ src/main/java/scs/planus/domain/group/dto/GroupDetailUpdateRequestDto.java
@@ -15,5 +15,5 @@ public class GroupDetailUpdateRequestDto {
 
     @Max(value = 50, message = "[request] 그룹 제한 인원은 50명 이하로 입력해 주세요.")
     @Min(value = 2, message = "[request] 그룹 제한 인원은 2명 이상으로 입력해 주세요.")
-    private Long limitCount;
+    private int limitCount;
 }
